+++ website_support/models/res_partner.py
@@ -5,11 +5,11 @@ class ResPartnerTicket(models.Model):
 
     _inherit = "res.partner"
 
-    support_ticket_ids = fields.One2many('website.support.ticket', 'partner_id', string='Tickets')
+    support_ticket_ids = fields.One2many('website.support.ticket', 'partner_id', string="Tickets")
     support_ticket_count = fields.Integer(compute="_count_support_tickets", string="Ticket Count")
     new_support_ticket_count = fields.Integer(compute="_count_new_support_tickets", string="New Ticket Count")
     support_ticket_string = fields.Char(compute="_compute_support_ticket_string", string="Support Ticket String")
-    stp_ids = fields.Many2many('res.partner', 'stp_res_partner_rel', 'stp_p1_id', 'stp_p2_id', string="Support Ticket Access Accounts", help="Can view support tickets from other contacts")
+    stp_ids = fields.Many2many('res.partner', 'stp_res_partner_rel', 'stp_p1_id', 'stp_p2_id', string="Support Ticket Access Accounts", help="(DEPRICATED use departments) Can view support tickets from other contacts")
     sla_id = fields.Many2one('website.support.sla', string="SLA")
     dedicated_support_user_id = fields.Many2one('res.users', string="Dedicated Support User")