+++ website_support/models/website_support_ticket.py
@@ -97,7 +97,8 @@ class WebsiteSupportTicket(models.Model):
     sla_timer = fields.Float(string="SLA Time Remaining")
     sla_timer_format = fields.Char(string="SLA Timer Format", compute="_compute_sla_timer_format")
     sla_active = fields.Boolean(string="SLA Active")
-    sla_response_category_id = fields.Many2one('website.support.sla.response', string="SLA Response Category")
+    sla_response_category_id = fields.Many2one('website.support.sla.response', string="(DEPRICATED) SLA Response Category")
+    sla_rule_id = fields.Many2one('website.support.sla.rule', string="SLA Rule")
     sla_alert_ids = fields.Many2many('website.support.sla.alert', string="SLA Alerts",
                                      help="Keep record of SLA alerts sent so we do not resend them")
 
@@ -114,10 +115,36 @@ class WebsiteSupportTicket(models.Model):
         for active_sla_ticket in self.env['website.support.ticket'].search([
             ('sla_active','=',True),
             ('sla_id','!=',False),
-            ('sla_response_category_id','!=',False)
+            '|',
+            ('sla_response_category_id','!=',False),
+            ('sla_rule_id','!=',False)
         ]):
 
             # If we only countdown during busines hours
+            if active_sla_ticket.sla_rule_id.countdown_condition == 'business_only':
+                # Check if the current time aligns with a timeslot in the settings,
+                # setting has to be set for business_only or UserError occurs
+                setting_business_hours_id = self.env['ir.default'].get('website.support.settings', 'business_hours_id')
+                current_hour = datetime.datetime.now().hour
+                current_minute = datetime.datetime.now().minute / 60
+                current_hour_float = current_hour + current_minute
+                day_of_week = datetime.datetime.now().weekday()
+                during_work_hours = self.env['resource.calendar.attendance'].search([('calendar_id','=', setting_business_hours_id), ('dayofweek','=',day_of_week), ('hour_from','<',current_hour_float), ('hour_to','>',current_hour_float)])
+
+                # If holiday module is installed take into consideration
+                holiday_module = self.env['ir.module.module'].search([('name','=','hr_public_holidays'), ('state','=','installed')])
+                if holiday_module:
+                    holiday_today = self.env['hr.holidays.public.line'].search([('date','=',datetime.datetime.now().date())])
+                    if holiday_today:
+                        during_work_hours = False
+
+                if during_work_hours:
+                    active_sla_ticket.sla_timer -= 1/60
+            elif active_sla_ticket.sla_rule_id.countdown_condition == '24_hour':
+                #Countdown even if the business hours setting is not set
+                active_sla_ticket.sla_timer -= 1/60
+                
+            #(DEPRICATED use sla_rule_id) If we only countdown during busines hours
             if active_sla_ticket.sla_response_category_id.countdown_condition == 'business_only':
                 # Check if the current time aligns with a timeslot in the settings,
                 # setting has to be set for business_only or UserError occurs
@@ -137,7 +164,7 @@ class WebsiteSupportTicket(models.Model):
 
                 if during_work_hours:
                     active_sla_ticket.sla_timer -= 1/60
-            else:
+            elif active_sla_ticket.sla_response_category_id.countdown_condition == '24_hour':
                 #Countdown even if the business hours setting is not set
                 active_sla_ticket.sla_timer -= 1/60
 
@@ -335,7 +362,35 @@ class WebsiteSupportTicket(models.Model):
 
         #Check if this contact has a SLA assigned
         if new_id.partner_id.sla_id:
-            #Check if this category has a SLA response time
+            
+            #Go through all rules starting from the lowest response time
+            for sla_rule in new_id.partner_id.sla_id.rule_ids:
+                #All conditions have to match
+                _logger.error(sla_rule.name)
+                all_true = True
+                for sla_rule_con in sla_rule.condition_ids:
+                    _logger.error("rule type:" + str(sla_rule_con.type))
+                    _logger.error("ticket category: " + str(new_id.category.name))
+                    _logger.error("rule category: " + str(sla_rule_con.category_id.name))
+                    _logger.error("ticket sub category: " + str(new_id.sub_category_id.name))
+                    _logger.error("rule sub category: " + str(sla_rule_con.subcategory_id.name))
+                    _logger.error("ticket priority: " + str(new_id.priority_id.name))
+                    _logger.error("rule priority: " + str(sla_rule_con.priority_id.name))
+                    if sla_rule_con.type == "category" and new_id.category.id != sla_rule_con.category_id.id:
+                        all_true = False
+                    elif sla_rule_con.type == "subcategory" and new_id.sub_category_id.id != sla_rule_con.subcategory_id.id:
+                        all_true = False
+                    elif sla_rule_con.type == "priority" and new_id.priority_id.id != sla_rule_con.priority_id.id:
+                        all_true = False
+                
+                if all_true:
+                    new_id.sla_id = new_id.partner_id.sla_id.id
+                    new_id.sla_active = True
+                    new_id.sla_timer = sla_rule.response_time
+                    new_id.sla_rule_id = sla_rule.id
+                    break
+
+            #(DEPRICATED) Check if this category has a SLA response time
             category_response = self.env['website.support.sla.response'].search([('vsa_id','=',new_id.partner_id.sla_id.id), ('category_id','=',new_id.category.id)])
             if category_response:
                 new_id.sla_id = new_id.partner_id.sla_id.id